+++ bot/utils.py
@@ -25,3 +25,13 @@ def check_date_format(date_string):
 def check_month_format(month_string):
     pattern = r"^(Jan|Feb|Mar|Apr|May|Jun|Jul|Aug|Sep|Oct|Nov|Dec)$"
     return bool(re.fullmatch(pattern, month_string, re.IGNORECASE))
+
+
+def sanitize_error_message(error_message):
+    # Regular expression to match URLs
+    url_pattern = r'https?:\/\/(?:www\.|(?!www))[a-zA-Z0-9][a-zA-Z0-9-]+[a-zA-Z0-9]\.[^\s]{2,}|www\.[a-zA-Z0-9][a-zA-Z0-9-]+[a-zA-Z0-9]\.[^\s]{2,}|https?:\/\/(?:www\.|(?!www))[a-zA-Z0-9]+\.[^\s]{2,}|www\.[a-zA-Z0-9]+\.[^\s]{2,}'
+    
+    # Replace URLs with "<url here>"
+    sanitized_message = re.sub(url_pattern, '<url here>', error_message)
+    
+    return sanitized_message